+++ build.gradle
@@ -1,13 +1,6 @@
 buildscript {
     ext {
         kotlin_version = '1.5.0'
-        filament_version = '1.12.0'
-        arcore_version = '1.26.0'
-        appcompat_version = '1.3.1'
-        corektx_version = '1.6.0'
-        lifecycle_version = '2.3.1'
-        material_version = '1.4.0'
-        okhttp_version = '3.12.13'
     }
     repositories {
         mavenCentral()
