+++ core/build.gradle
@@ -36,10 +36,13 @@ android {
 }
 
 dependencies {
+    def filament_version = '1.12.0'
+    def arcore_version = '1.26.0'
+
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlin_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$kotlin_version"
-
+    
     api "com.google.android.filament:filament-android:$filament_version"
     api "com.google.android.filament:gltfio-android:$filament_version"
     api "com.google.android.filament:filament-utils-android:$filament_version"
@@ -47,16 +50,14 @@ dependencies {
 
     api "com.google.ar:core:$arcore_version"
 
-    implementation "androidx.appcompat:appcompat:$appcompat_version"
-    implementation "androidx.core:core-ktx:$corektx_version"
-    implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_version"
-    implementation "com.google.android.material:material:$material_version"
+    implementation "androidx.appcompat:appcompat:1.3.1"
+    implementation "androidx.core:core-ktx:1.6.0"
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.3.1"
+    implementation "com.google.android.material:material:1.4.0"
 
     implementation "com.github.kittinunf.fuel:fuel:2.3.1"
     implementation "com.github.kittinunf.fuel:fuel-android:2.3.1"
     implementation "com.github.kittinunf.fuel:fuel-coroutines:2.3.1"
-
-//    implementation "com.squareup.okhttp3:okhttp:$okhttp_version"
 }
 
 if (project.hasProperty('mavenCentralUsername') && project.hasProperty('mavenCentralPassword')) {
