+++ core/src/main/java/com/gorisse/thomas/sceneform/ArSceneView.kt
@@ -1,49 +1,71 @@
 package com.gorisse.thomas.sceneform
 
 import com.google.ar.sceneform.ArSceneView
-import com.gorisse.thomas.sceneform.environment.Environment
-import com.gorisse.thomas.sceneform.filament.Light
-import com.gorisse.thomas.sceneform.filament.clone
-import com.gorisse.thomas.sceneform.filament.destroy
+import com.gorisse.thomas.sceneform.light.*
+
 
 /**
- * ### The environment that is estimated by AR Core to render the scene.
+ * ### ARCore light estimation configuration
  *
- * Environment handles a reflections, indirect lighting and skybox
- */
-var ArSceneView.estimatedEnvironment: Environment?
-    get() = _estimatedEnvironment
-    internal set(value) {
-        renderer?.setEnvironment(value ?: environment)
-        _estimatedEnvironment?.destroy()
-        _estimatedEnvironment = value
-    }
-
-/**
- * ### The main light that is estimated by AR Core to render the scene.
+ * ARCore estimate lighting to provide directional light, ambient spherical harmonics,
+ * and reflection cubemap estimation
  *
- * Ar Core will estimate the direction, the intensity and the color of the light
+ * Light bounces off of surfaces differently depending on whether the surface has specular
+ * (highly reflective) or diffuse (not reflective) properties.
+ * For example, a metallic ball will be highly specular and reflect its environment, while
+ * another ball painted a dull matte gray will be diffuse. Most real-world objects have a
+ * combination of these properties — think of a scuffed-up bowling ball or a well-used credit
+ * card.
+ *
+ * Reflective surfaces also pick up colors from the ambient environment. The coloring of an
+ * object can be directly affected by the coloring of its environment. For example, a white ball
+ * in a blue room will take on a bluish hue.
+ *
+ * The main directional light API calculates the direction and intensity of the scene's
+ * main light source. This information allows virtual objects in your scene to show reasonably
+ * positioned specular highlights, and to cast shadows in a direction consistent with other
+ * visible real objects.
+ *
+ * @see LightEstimationConfig.REALISTIC
+ * @see LightEstimationConfig.SPECTACULAR
+ * @see LightEstimationConfig.AMBIENT_INTENSITY
  */
-var ArSceneView.estimatedMainLight: Light?
-    get() = _estimatedMainLight
-    internal set(value) {
-        renderer?.setMainDirectionalLight(value ?: mainDirectionalLight)
-        _estimatedMainLight?.destroy()
-        _estimatedMainLight = value
-    }
-
-// TODO : Move to internal when ArSceneView is fully Kotlined
-var ArSceneView.estimatedMainLightInfluence: (Light.() -> Unit)?
-    get() = _estimatedMainLightInfluence
-    internal set(value) {
-        _estimatedMainLightInfluence = value
-        if (value != null) {
-            val estimatedMainLight = estimatedMainLight
-                ?: mainDirectionalLight?.clone()?.also {
-                    estimatedMainLight = it
-                }
-            estimatedMainLight?.apply(value)
-        } else {
-            estimatedMainLight = null
+var ArSceneView.lightEstimationConfig: LightEstimationConfig get() = _lightEstimationConfig
+    set(value) {
+        if(_lightEstimationConfig != value) {
+            if(value.mode != sessionConfig?.lightEstimationMode) {
+               setSessionConfig(sessionConfig?.apply {
+                   lightEstimationMode = value.mode
+               }, true)
+            }
+            estimatedEnvironmentLights = null
+            _lightEstimationConfig = value
         }
+    }
+
+/**
+ * ### The environment and main light that are estimated by AR Core to render the scene.
+ *
+ * - Environment handles a reflections, indirect lighting and skybox.
+ *
+ * - ARCore will estimate the direction, the intensity and the color of the light
+ */
+var ArSceneView.estimatedEnvironmentLights: EnvironmentLightsEstimate?
+    get() = _estimatedEnvironmentLights
+    internal set(value) {
+        val environment = value?.environment ?: environment
+        if(renderer?.getEnvironment() != environment) {
+            if(_estimatedEnvironmentLights?.environment != environment) {
+                _estimatedEnvironmentLights?.environment?.destroy()
+            }
+            renderer?.setEnvironment(environment)
+        }
+        val mainLight = value?.mainLight ?: mainLight
+        if(renderer?.getMainLight() != mainLight) {
+            if(_estimatedEnvironmentLights?.mainLight != mainLight) {
+                _estimatedEnvironmentLights?.mainLight?.destroy()
+            }
+            renderer?.setMainLight(mainLight)
+        }
+        _estimatedEnvironmentLights= value
     }
