--- core/src/main/java/com/gorisse/thomas/sceneform/arcore/Anchor.kt
@@ -1,7 +1,0 @@
-package com.gorisse.thomas.sceneform.arcore
-
-import com.google.ar.core.Anchor
-
-fun Anchor.destroy() {
-    detach()
-}
