--- core/src/main/java/com/gorisse/thomas/sceneform/filament/Filament.kt
@@ -1,53 +1,0 @@
-package com.gorisse.thomas.sceneform.filament
-
-import com.google.android.filament.EntityManager
-import com.google.android.filament.gltfio.AssetLoader
-import com.google.android.filament.gltfio.ResourceLoader
-import com.google.android.filament.gltfio.UbershaderLoader
-import com.google.android.filament.utils.Float3
-import com.gorisse.thomas.sceneform.environment.IBLPrefilter
-import com.google.ar.sceneform.rendering.EngineInstance
-
-object Filament {
-
-    @JvmStatic
-    val engine = EngineInstance.getEngine().filamentEngine
-
-    @JvmStatic
-    val entityManager
-        get() = EntityManager.get()
-
-    val uberShaderLoader by lazy { UbershaderLoader(engine) }
-
-    @JvmStatic
-    val assetLoader by lazy {
-        AssetLoader(engine, uberShaderLoader, entityManager)
-    }
-
-    val transformManager get() = engine.transformManager
-
-    val resourceLoader by lazy { ResourceLoader(engine, true, false) }
-
-    val lightManager get() = engine.lightManager
-
-    val iblPrefilter by lazy { IBLPrefilter(engine) }
-}
-
-fun Float3.toFloatArray() = this.let { (x, y, z) -> floatArrayOf(x, y, z) }
-fun FloatArray.toFloat3() = this.let { (x, y, z) -> Float3(x, y, z) }
-
-typealias Position = Float3
-
-fun FloatArray.toPosition() = this.let { (x, y, z) -> Position(x, y, z) }
-fun positionOf(x: Float = 0.0f, y: Float = 0.0f, z: Float = 0.0f) = Position(x, y, z)
-
-typealias Direction = Float3
-
-fun FloatArray.toDirection() = this.let { (x, y, z) -> Direction(x, y, z) }
-fun directionOf(x: Float = 0.0f, y: Float = 0.0f, z: Float = 0.0f) = Direction(x, y, z)
-
-typealias Color = Float3
-
-fun FloatArray.toColor() = this.let { (r, g, b) -> Color(r, g, b) }
-fun colorOf(r: Float = 0.0f, g: Float = 0.0f, b: Float = 0.0f) = Color(r, g, b)
-
