--- core/src/main/java/com/gorisse/thomas/sceneform/filament/Texture.kt
@@ -1,18 +1,0 @@
-package com.gorisse.thomas.sceneform.filament
-
-import com.google.android.filament.Skybox
-import com.google.android.filament.Texture
-
-fun <R> Texture.use(block: (Texture) -> R): R = block(this).also { destroy() }
-
-/**
- * @see Skybox.Builder.build
- */
-fun Texture.Builder.build(): Texture = build(Filament.engine)
-
-/**
- * Destroys a Texture and frees all its associated resources.
- */
-fun Texture.destroy() {
-    Filament.engine.destroyTexture(this)
-}
