+++ core/src/main/java/com/gorisse/thomas/sceneform/scene/Anchor.kt
@@ -1,0 +1,7 @@
+package com.gorisse.thomas.sceneform.scene
+
+import com.google.ar.core.Anchor
+
+fun Anchor.destroy() {
+    detach()
+}
