+++ docs/assets/environments/index.json
@@ -7,5 +7,9 @@
   {
     "name": "Evening Meadow (HDR)",
     "hdrUrl": "evening_meadow_2k.hdr"
+  },
+  {
+    "name": "Flower Road (HDR)",
+    "hdrUrl": "flower_road_2k.hdr"
   }
 ]
