+++ samples/environment-lights/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.gorisse.thomas.ar.environmentlights">
+    package="com.google.ar.sceneform.samples.environmentlights">
     <!-- "AR Required" apps must declare minSdkVersion ≥ 24.
          "AR Optional" apps must declare minSdkVersion ≥ 14 -->
     <!-- Always needed for AR. -->
@@ -49,6 +49,7 @@
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
         </activity>
     </application>
 </manifest>
