+++ samples/environment-lights/src/main/res/values/themes.xml
@@ -4,10 +4,10 @@
         <item name="colorPrimary">#FF4b8eff</item>
         <item name="colorPrimaryDark">@android:color/transparent</item>
         <item name="colorAccent">#FF4b8eff</item>
-        <item name="android:textColor">@android:color/white</item>
-        <item name="colorButtonNormal">#FF4b8eff</item>
         <item name="android:statusBarColor">@android:color/transparent</item>
         <item name="android:windowTranslucentNavigation">true</item>
+        <!-- Text -->
+        <item name="android:textColor">@android:color/white</item>
     </style>
 
     <style name="AppTheme.Toolbar" parent="ThemeOverlay.AppCompat.Dark.ActionBar">
